Conflicts:
	src/main/java/com/soapboxrace/core/bo/InviteTicketBO.java

+++ src/main/java/com/soapboxrace/core/bo/TokenSessionBO.java
@@ -69,12 +69,14 @@ public class TokenSessionBO {
 		LoginStatusVO loginStatusVO = new LoginStatusVO(0L, "", false);
 		if (email != null && !email.isEmpty() && password != null && !password.isEmpty()) {
 			UserEntity userEntity = userDAO.findByEmail(email);
-			if (password.equals(userEntity.getPassword())) {
-				Long userId = userEntity.getId();
-				deleteByUserId(userId);
-				String randomUUID = createToken(userId);
-				loginStatusVO = new LoginStatusVO(userId, randomUUID, true);
-				return loginStatusVO;
+			if (userEntity != null) {
+				if (password.equals(userEntity.getPassword())) {
+					Long userId = userEntity.getId();
+					deleteByUserId(userId);
+					String randomUUID = createToken(userId);
+					loginStatusVO = new LoginStatusVO(userId, randomUUID, true);
+					return loginStatusVO;
+				}
 			}
 		}
 		loginStatusVO.setDescription("LOGIN ERROR");
@@ -95,6 +97,17 @@ public class TokenSessionBO {
 	public String getActiveRelayCryptoTicket(String securityToken) {
 		TokenSessionEntity tokenSessionEntity = tokenDAO.findById(securityToken);
 		return tokenSessionEntity.getRelayCryptoTicket();
+	}
+
+	public Long getActiveLobbyId(String securityToken) {
+		TokenSessionEntity tokenSessionEntity = tokenDAO.findById(securityToken);
+		return tokenSessionEntity.getActiveLobbyId();
+	}
+
+	public void setActiveLobbyId(String securityToken, Long lobbyId) {
+		TokenSessionEntity tokenSessionEntity = tokenDAO.findById(securityToken);
+		tokenSessionEntity.setActiveLobbyId(lobbyId);
+		tokenDAO.update(tokenSessionEntity);
 	}
 
 }
